+++ app/routes/app.tsx
@@ -38,6 +38,7 @@ export default function App() {
         <Link to="/app/order-rules">Order rules</Link>
         <Link to="/app/credit">Credit</Link>
         <Link to="/app/payments">Payments</Link>
+        <Link to="/app/tax">Tax &amp; VAT</Link>
         <Link to="/app/accounting">Accounting</Link>
         <Link to="/app/settings">Settings</Link>
       </NavMenu>